pyproject.toml
[project]
name = "redboxhal"
version = "0.1.0"
description = "An open source, cross platform, Hardware Abstraction Layer for Redbox kiosks"
authors = [{ name = "Puyodead1", email = "puyodead@proton.me" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pyserial = "^3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150
